+++ packages/stdnum-linux-x64-gnu/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @stll/stdnum-linux-x64-gnu
 
+## 2.3.2
+
 ## 2.3.1
 
 ## 2.3.0